+++ whatshap/haplotag.py
@@ -48,6 +48,11 @@ def add_arguments(parser):
 	arg('--reference', '-r', metavar='FASTA',
 		help='Reference file. Provide this to detect alleles through re-alignment. '
 			'If no index (.fai) exists, it will be created')
+	arg('--ignore-linked-read', default=False, action='store_true',
+		help='Ignore linkage information stored in BX tags of the reads.')
+	arg('--linked-read-distance-cutoff', '-d', metavar='LINKEDREADDISTANCE', default=50000, type=int,
+		help='Assume reads with identical BX tags belong to different read clouds if their '
+			'distance is larger than LINKEDREADDISTANCE (default: %(default)s).')
 	arg('variant_file', metavar='VCF', help='VCF file with phased variants (can be gzip-compressed)')
 	arg('alignment_file', metavar='ALIGNMENTS',
 		help='File (BAM/CRAM) with read alignments to be tagged by haplotype')
@@ -61,7 +66,9 @@ def md5_of(filename):
 	return md5(open(filename,'rb').read()).hexdigest()
 
 
-def run_haplotag(variant_file, alignment_file, output=None, reference=None):
+def run_haplotag(variant_file, alignment_file, output=None, reference=None, ignore_linked_read=False,
+	linked_read_distance_cutoff=50000):
+
 	timers = StageTimer()
 	timers.start('overall')
 
@@ -111,7 +118,7 @@ def run_haplotag(variant_file, alignment_file, output=None, reference=None):
 		# TODO: convince pysam to allow @HS header line
 		header = bam_reader.header.to_dict()
 		command_line = ' '.join(['whatshap'] + sys.argv[1:])
-		PG_entry = { 'PN':'whatshap', 'VN':__version__, 'CL':command_line, 'm5': md5_of(variant_file)}
+		PG_entry = {'ID':'whatshap', 'PN':'whatshap', 'VN':__version__, 'CL':command_line, 'm5': md5_of(variant_file)}
 		if 'PG' in header:
 			header['PG'].append(PG_entry)
 		else:
@@ -128,6 +135,10 @@ def run_haplotag(variant_file, alignment_file, output=None, reference=None):
 		n_alignments = 0
 		n_tagged = 0
 		n_multiple_phase_sets = 0
+
+		# map BX tag to assigned haplotype
+		BX_tag_to_haplotype = defaultdict(list)
+
 		for alignment in bam_reader:
 			n_alignments += 1
 			alignment.set_tag('HP', value=None)
@@ -138,6 +149,7 @@ def run_haplotag(variant_file, alignment_file, output=None, reference=None):
 				if chromosome_id != alignment.reference_id:
 					chromosome_id = alignment.reference_id
 					chromosome_name = alignment.reference_name
+					BX_tag_to_haplotype = defaultdict(list)
 					logger.info('Processing alignments on chromosome %s', chromosome_name)
 					if chromosome_name in skipped_vcf_chromosomes:
 						logger.error('Chromosome records in alignment file and VCF are sorted differently.')
@@ -160,23 +172,48 @@ def run_haplotag(variant_file, alignment_file, output=None, reference=None):
 						for sample in samples:
 							genotypes = variant_table.genotypes_of(sample)
 							phases = variant_table.phases_of(sample)
-							variantpos_to_phaseset = {
-								v.position:int(phases[i].block_id) for i,v in enumerate(variant_table.variants) if phases[i] is not None
+							variantpos_to_phaseinfo = {
+								v.position:(int(phases[i].block_id),phases[i].phase) for i,v in enumerate(variant_table.variants) if phases[i] is not None
 							}
 							variants = [
 								v for v, gt, phase in zip(variant_table.variants, genotypes, phases) if gt == 1 and phase is not None
 							]
 							read_set = read_reads(readset_reader, chromosome_name, variants, sample, fasta)
+
+							# map tag --> set of reads
+							BX_tag_to_readlist = defaultdict(list)
 							for read in read_set:
+								if read.has_BX_tag():
+									BX_tag_to_readlist[read.BX_tag].append(read)
+							# all reads processed so far
+							processed_reads = set()
+							for read in read_set:
+								if read.name in processed_reads:
+									continue
 								# mapping: phaseset --> phred scaled difference between costs of assigning reads to haplotype 0 or 1
 								haplotype_costs = defaultdict(int)
-								for v in read:
-									assert v.allele in [0,1]
-									phaseset = variantpos_to_phaseset[v.position]
-									if v.allele == 0:
-										haplotype_costs[phaseset] += v.quality
-									else:
-										haplotype_costs[phaseset] -= v.quality
+								reads_to_consider = set()
+
+								processed_reads.add(read.name)
+								reads_to_consider.add(read)
+
+								# reads with same BX tag need to be considered too (unless --ignore-linked-read is set)
+								if read.has_BX_tag() and not ignore_linked_read:
+									for r in BX_tag_to_readlist[read.BX_tag]:
+										if not r.name in processed_reads:
+											# only select reads close to current one
+											if abs(read.reference_start - r.reference_start) <= linked_read_distance_cutoff:
+												reads_to_consider.add(r)
+								for r in reads_to_consider:
+									processed_reads.add(r.name)
+									for v in r:
+										assert v.allele in [0,1]
+										phaseset, allele = variantpos_to_phaseinfo[v.position]
+										if v.allele == allele:
+											haplotype_costs[phaseset] += v.quality
+										else:
+											haplotype_costs[phaseset] -= v.quality
+
 								l = list(haplotype_costs.items())
 								l.sort(key=lambda t:-abs(t[1]))
 								#logger.info('Read %s: %s', read.name, str(l))
@@ -186,8 +223,10 @@ def run_haplotag(variant_file, alignment_file, output=None, reference=None):
 									phaseset, quality = l[0]
 									if quality != 0:
 										haplotype = 0 if quality > 0 else 1
-										read_to_haplotype[read.name] = (haplotype, abs(quality), phaseset)
-										#logger.debug('Assigned read %s to haplotype %d with a quality of %d based on %d covered variants', read.name, haplotype, quality, len(read))
+										BX_tag_to_haplotype[read.BX_tag].append((read.reference_start, haplotype, phaseset))
+										for r in reads_to_consider:
+											read_to_haplotype[r.name] = (haplotype, abs(quality), phaseset)
+											logger.debug('Assigned read %s to haplotype %d with a quality of %d based on %d covered variants', r.name, haplotype, quality, len(r))
 
 				# Only attempt to assign phase of neither secondary nor supplementary
 				if (not alignment.is_secondary) and (alignment.flag & 2048 == 0):
@@ -198,7 +237,15 @@ def run_haplotag(variant_file, alignment_file, output=None, reference=None):
 						alignment.set_tag('PS', phaseset)
 						n_tagged += 1
 					except KeyError:
-						pass
+						# check if reads with same tag have been assigned
+						if alignment.has_tag('BX'):
+							read_clouds = BX_tag_to_haplotype[alignment.get_tag('BX')]
+							for (reference_start, haplotype, phaseset) in read_clouds:
+								if abs(reference_start - alignment.reference_start) <= linked_read_distance_cutoff:
+									alignment.set_tag('HP', haplotype + 1)
+									alignment.set_tag('PS', phaseset)
+									n_tagged += 1
+									break
 			bam_writer.write(alignment)
 			if n_alignments % 100000 == 0:
 				logger.info('Processed %d alignment records.', n_alignments)